pyproject.toml
[tool.poetry]
name = "goit-python-web-hw-10"
version = "0.1.0"
description = ""
authors = ["lexxai <lexxai@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"
pymongo = "^4.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
